beta version 0.5

+++ tetris/include/functional/settings.h
@@ -2,6 +2,7 @@
 
 namespace Settings {
 
-    constexpr static size_t _colNum = 10;
-    constexpr static size_t _rowNum = 20;
+    constexpr static int _colNum = 10;
+    constexpr static int _rowNum = 20;
+
 }
